moved solution to general as it is something expected from all model fixers
