+++ src/hvl/projectparmorel/general/ModelFixer.java
@@ -4,7 +4,6 @@ import java.io.File;
 import java.util.List;
 
 import hvl.projectparmorel.exceptions.NoErrorsInModelException;
-import hvl.projectparmorel.modelrepair.Solution;
 import hvl.projectparmorel.reward.PreferenceOption;
 
 public interface ModelFixer {
